Add notebooks w/preliminary analysis and effect on active editors
The preliminary analysis notebook examines how to identify micro-
contributions and establishes a timeline of releases on Wikidata
and Commons. Then looks at some overall stats (e.g. number of
contributions), and whether data in account age is meaningful.

The other notebook looks at the extent of micro-contributions
done by active editors, particularly with regards to the effect
on the measurement.